requirements.txt
Django==5.2.12
Pillow==11.2.1
whitenoise==6.9.0
psycopg[binary]==3.2.9
python-dotenv==1.1.0
redis==5.0.1
django-redis==5.4.0
sentry-sdk==1.40.0
celery==5.3.4
django-celery-results==2.6.0
gunicorn==21.2.0
dj-database-url==2.1.0
